requirements.txt
yapsy == 1.12.2
python_dateutil == 2.6.0
requests
PyJWT >=1.7.1
cryptography
sqlalchemy
psycopg2-binary
pyyaml
fabric-message-bus